fix some CodeQL warnings

+++ client/src/cmdhfmfu.c
@@ -2104,7 +2104,7 @@ void printMFUdumpEx(mfu_dump_t *card, uint16_t pages, uint8_t startpage) {
         PrintAndLogEx(INFO, "DYNAMIC LOCK: %s", sprint_hex(lockbytes_dyn, 3));
     }
 
-    for (uint8_t i = 0; i < pages; ++i) {
+    for (uint16_t i = 0; i < pages; ++i) {
         if (i < 3) {
             PrintAndLogEx(INFO, "%3d/0x%02X | %s|   | %s", i + startpage, i + startpage, sprint_hex(data + i * 4, 4), sprint_ascii(data + i * 4, 4));
             continue;
