+++ client/src/cmdlfparadox.c
@@ -102,7 +102,7 @@ int demodParadox(bool verbose) {
     uint8_t error = 0;
 
     // Remove manchester encoding from FSK bits, skip pre
-    for (uint8_t i = idx + PARADOX_PREAMBLE_LEN; i < (idx + 96); i += 2) {
+    for (uint32_t i = idx + PARADOX_PREAMBLE_LEN; i < (idx + 96); i += 2) {
 
         // not manchester data
         if (bits[i] == bits[i + 1]) {
