+++ client/src/util.c
@@ -280,8 +280,8 @@ void print_blocks(uint32_t *data, size_t len) {
     if (!data) {
         PrintAndLogEx(ERR, "..empty data");
     } else {
-        for (uint8_t i = 0; i < len; i++)
-            PrintAndLogEx(SUCCESS, " %02d | %08X", i, data[i]);
+        for (size_t i = 0; i < len; i++)
+            PrintAndLogEx(SUCCESS, " %02zd | %08X", i, data[i]);
     }
 }
